Rework lexical closure and the symbol table
To support mutability of the symbol table, especially for a nested
block to modify a value in the parent (lexical) block, we need some
sort of scope chain representation. Due to closures we may also have
to retain old scope chains which in practice turns the chains into a
scope tree, rooted in the initial scope.

This patch introduces a simple scope tree, with no garbage collection
whatsoever. As the interpreter walks over the code it enters and
leaves scopes and the symbol table keeps track of the current scope
chain. Functions are bound to the current scope chain such that
entering/calling them will restore the scope chain to how it was when
the function was defined.

We may then remove some ugly crutches from the interpreter and move
some complexity into the symbol table.