Implement member functions aka methods
Introduce bound expressions as intermediate expressions in the
interpreter.

E.g. given

> let arr = array[int](1, 2, 3)
[1, 2, 3]

let's query for the length of the array:

> arr.len()
3

This looks up the structure corresponding to the object (hard-coded
for now; there is only the Array structure) and looks up the function
'len'.

Then it creates a tuple (the binding) of the instance and the function:

(fn len() .. <-> [1,2,3])

When this binding is evaluated, the argument list is modified to
prepend the instance first and the 'len' function sees the array
instance as the first argument. So arr.len() is basically translated
to:

Array.len(arr)

Note however that the type-checker will now allow the expression
'Array.len(arr)', only the interpreter may create this intermediate
form. The type checker things the Array.len function takes exactly 0
arguments.

> Array.len(arr)
Error: Type resolve error: Type mismatch: argument mismatch: "expected 0 arguments, got 1"

This leaves the question what to do about:

> Array.len()
Error: Interpreter error: Missing or wrong number of arguments

This is an interpreter error when it should be a type checker error.
The type checker thus lets it through! Will have to adress that later
somehow. Will probably have to amend the ResolvedType for the len
function with the binding requirement.